+++ coding-dojo-huawei/src/test/java/com/uj/study/prime/factors/PrimeFactorTest.java
@@ -16,29 +16,29 @@ public class PrimeFactorTest {
     @Test
     void calculatePrimeListWithinInputNumber() {
         int input = 2;
-        List<Integer> primes = PrimeFactor.getPrimeList(input);
+        List<Long> primes = PrimeFactor.getPrimeList(input);
         assertThat(primes.size(), is(1));
-        assertThat(primes.get(0), is(2));
+        assertThat(primes.get(0), is(2L));
 
-        primes = PrimeFactor.getPrimeList(3);
+        primes = PrimeFactor.getPrimeList(3L);
         assertThat(primes.size(), is(2));
 
-        primes = PrimeFactor.getPrimeList(4);
+        primes = PrimeFactor.getPrimeList(4L);
         assertThat(primes.size(), is(2));
 
-        primes = PrimeFactor.getPrimeList(5);
+        primes = PrimeFactor.getPrimeList(5L);
         assertThat(primes.size(), is(3));
 
-        primes = PrimeFactor.getPrimeList(180);
+        primes = PrimeFactor.getPrimeList(180L);
         assertThat(primes.size(), is(41));
     }
 
     @Test
     void calcualtePrimeFactor() {
-        assertThat(PrimeFactor.getPrimeFactor(12), is(Arrays.asList(2,2,3)));
-        assertThat(PrimeFactor.getPrimeFactor(147), is(Arrays.asList(3,7,7)));
-        assertThat(PrimeFactor.getPrimeFactor(17), is(Arrays.asList(17)));
-        assertThat(PrimeFactor.getPrimeFactor(180), is(Arrays.asList(2, 2, 3, 3, 5)));
+        assertThat(PrimeFactor.getPrimeFactor(12), is(Arrays.asList(2L, 2L, 3L)));
+        assertThat(PrimeFactor.getPrimeFactor(147), is(Arrays.asList(3L, 7L ,7L)));
+        assertThat(PrimeFactor.getPrimeFactor(17), is(Arrays.asList(17L)));
+        assertThat(PrimeFactor.getPrimeFactor(180), is(Arrays.asList(2L, 2L, 3L, 3L, 5L)));
     }
 
     public void test3(int n){
